Addressing some of the issues from the PR
- `tapeDataOut()` calls `scsiFinishWrite()` after calls to `scsiStartRead()`. This fails at least on RP2040, but I think in addition to wrong function there is something else wrong with the logic here. Perhaps synchronous `scsiRead()` could be used for now, even though it has worse performance.
 Switched to SCSI Read

- Read operations appear limited to max 64 kB requests.
Not sure if this has been addressed

- I think the length calculations for fixed length writes is wrong. According to SCSI spec, the argument should be number of blocks but it appears to be treated as a number of bytes.
for a fixed length switched to multiplying length times bytes_per_sector

- I'm not sure if Seek and ReadPosition work correctly for TAP files - maybe they do.
haven't addressed this

- `img.tape_pos` is 32-bit, which limits file to 4 GB when using the byte-based TAP format
Switched a 64-bit variable. It should be noted a record can only be 4GB
in length due to the SIMH record length being 32bit. But this should
accommodate multiple records totaling over 4GB.

An issue has been discoverd where on fixed reads, the whole record has
to be read at once instead of single blocks sequentially.
Line 403 of `src/ZuluSCSI_tape.cpp`

+++ src/ZuluSCSI_disk.h
@@ -60,7 +60,7 @@ struct image_config_t: public S2S_TargetCfg
     uint8_t ejectButton;
 
     // For tape drive emulation
-    uint32_t tape_pos; // current position in blocks
+    uint64_t tape_pos; // current position in blocks
     uint32_t tape_mark_index; // a direct relationship to the file in a multi image file tape 
     uint32_t tape_mark_count; // the number of marks
     uint32_t tape_mark_block_offset; // Sum of the the previous image file sizes at the current mark
